test(docx-reader): unify DOCX corpus snapshot harness into snapshots.rs
