+++ crates/docspec-docx-reader/build.rs
@@ -9,15 +9,20 @@ use std::path::{Path, PathBuf};
 
 fn main() -> Result<(), Box<dyn core::error::Error>> {
     println!("cargo::rustc-check-cfg=cfg(coverage)");
-    generate_pandoc_corpus_tests()?;
+    generate_corpus_tests("pandoc", "pandoc_corpus_tests.rs")?;
+    generate_corpus_tests("docspec", "docspec_corpus_tests.rs")?;
     Ok(())
 }
 
-fn generate_pandoc_corpus_tests() -> Result<(), Box<dyn core::error::Error>> {
+fn generate_corpus_tests(
+    corpus: &str,
+    output_file: &str,
+) -> Result<(), Box<dyn core::error::Error>> {
     let manifest_dir = PathBuf::from(env::var("CARGO_MANIFEST_DIR")?);
-    let fixture_dir = manifest_dir.join("../../tests/fixtures/docx/pandoc");
+    let fixture_dir = manifest_dir.join(format!("../../tests/fixtures/docx/{corpus}"));
     println!("cargo::rerun-if-changed={}", fixture_dir.display());
 
+    let corpus_literal = rust_string_literal(corpus);
     let mut generated = String::from("// @generated by crates/docspec-docx-reader/build.rs\n\n");
 
     if fixture_dir.try_exists()? {
@@ -31,13 +36,13 @@ fn generate_pandoc_corpus_tests() -> Result<(), Box<dyn core::error::Error>> {
             let fixture_literal = rust_string_literal(&fixture);
             writeln!(
                 &mut generated,
-                "pandoc_fixture_test!({test_name}, {fixture_literal});"
+                "#[test]\nfn {test_name}() {{\n    super::assert_fixture_snapshot({corpus_literal}, {fixture_literal});\n}}\n"
             )?;
         }
     }
 
     let out_dir = PathBuf::from(env::var("OUT_DIR")?);
-    fs::write(out_dir.join("pandoc_corpus_tests.rs"), generated)?;
+    fs::write(out_dir.join(output_file), generated)?;
     Ok(())
 }
 
@@ -77,6 +82,7 @@ fn test_name_for_fixture(file_name: &str) -> Result<String, Box<dyn core::error:
         "instrText_hyperlink" => "instr_text_hyperlink",
         "lists-compact" => "lists_compact",
         "nested_instrText" => "nested_instr_text",
+        "preformatted-boundaries" => "preformatted_boundaries",
         other => other,
     };
