+++ tests/fixtures/docx/pandoc/README.md
@@ -8,8 +8,8 @@ Git LFS. See `../../ATTRIBUTION.md` for license and source details.
 
 The corpus is copied from the sibling `documents` mirror at
 `documents/docx/pandoc/`. It intentionally preserves the upstream fixture names
-so `crates/docspec-docx-reader/tests/pandoc_corpus.rs` can discover every
-`.docx` file with a single `insta::glob!`.
+so the `pandoc` module in `crates/docspec-docx-reader/tests/snapshots.rs` can
+emit one test per `.docx` via `build.rs`.
 
 The fixtures cover headings, inline formatting, links, lists, tables, images,
 notes, metadata, tracked changes, structured document tags, and other WordprocessingML
@@ -38,6 +38,6 @@ To refresh from the sibling `documents` repo:
 rsync -a ../documents/documents/docx/pandoc/*.docx tests/fixtures/docx/pandoc/
 ```
 
-Then run `INSTA_UPDATE=unseen cargo test --test pandoc_corpus -p docspec-docx-reader`
+Then run `INSTA_UPDATE=unseen cargo test --test snapshots -p docspec-docx-reader`
 to generate any new snapshots, review them with `cargo insta review`, and commit.
 Update the **Imported** line in `../../ATTRIBUTION.md` when refreshing en masse.
